refactor(locale)!: use snake case for all locale data (#2910)
