+++ src/locales/base/system/directory_paths.ts
