+++ src/locales/en/science/chemical_element.ts
