+++ src/locales/eo/science/chemical_element.ts
