+++ src/locales/nb_NO/science/chemical_element.ts
