+++ src/locales/pl/science/chemical_element.ts
