+++ src/locales/zh_CN/science/chemical_element.ts
