+++ src/modules/science/index.ts
@@ -49,7 +49,7 @@ export class ScienceModule extends ModuleBase {
    */
   chemicalElement(): ChemicalElement {
     return this.faker.helpers.arrayElement(
-      this.faker.definitions.science.chemicalElement
+      this.faker.definitions.science.chemical_element
     );
   }
